Update dependencies (#456)
* Update phpstant to 0.12.87

* Update all dependencies

+++ src/Query/User/Model/Installs.php
@@ -34,7 +34,7 @@ final class Installs
 
     public function daysTotal(): int
     {
-        return (int) array_sum(array_map(fn (Day $day) => $day->installs(), $this->days));
+        return array_sum(array_map(fn (Day $day) => $day->installs(), $this->days));
     }
 
     public function total(): int
